Update application.properties for Swagger UI and Hibernate dialect
Added a new entry for the 'movement-event' API in the Swagger UI configuration by defining its name and URL. This allows for clearer documentation and access to the movement-event API through the Swagger interface. Additionally, specified the Hibernate dialect as H2Dialect to ensure proper configuration for H2 database compatibility during development. These changes enhance the API documentation and improve database handling in the application.