Add RestExceptionHandler to handle domain exceptions
This commit introduces a new class, RestExceptionHandler, which is responsible for mapping domain exceptions to HTTP responses in the API. By using the @RestControllerAdvice annotation, this handler allows centralized exception handling for specific exceptions, such as MovementInsufficientBalanceException. 

When this exception is thrown, the handler generates a detailed HTTP response that includes a 400 Bad Request status, a title indicating the nature of the error (Saldo no disponible), as well as the request URI and an error code (movement.insufficient_balance). This approach keeps controller code focused on handling successful paths, improving code maintainability and clarity.