Add unit tests for the Client model
Created a new test class `ClientTest` to validate the functionality of the `Client` model. This includes a test method that checks if the setters correctly populate the fields and ensures that the default state of the `active` field is set to true. The test verifies the following:
- The `clientCode`, `passwordHash`, and `active` fields are correctly assigned and retrieved.
- The `active` field defaults to true upon creation of the `Client` instance.
- The ability to set and retrieve the `active` status and associated `Person` object is confirmed.

This unit test is essential for maintaining the integrity of the `Client` model as we continue to develop the application.